requirements.txt
GitPython>=3.1.32
Pillow>=10.0.0
accelerate>=0.23.0

blendmodes>=2022
clean-fid>=0.1.35
diskcache>=5.6.1
einops>=0.7.0
facexlib>=0.3.0
fastapi>=0.104.1
gradio==3.41.2  # Keeping specific version as required
inflection>=0.5.1
jsonmerge>=1.9.2
kornia>=0.7.0
lark>=1.1.7
numpy>=1.24.0
omegaconf>=2.3.0
open-clip-torch>=2.20.0

piexif>=1.1.3
protobuf==3.20.0  # Keeping specific version as required
psutil>=5.9.0
pytorch_lightning>=2.1.0
requests>=2.31.0
resize-right>=0.0.2

safetensors>=0.4.0
scikit-image>=0.21.0
tomesd>=0.1.3
torch>=2.1.0
torchdiffeq>=0.2.3
torchsde>=0.2.6
transformers==4.30.2  # Keeping specific version as required
pillow-avif-plugin==1.4.3  # Keeping specific version as required

# Added Python version specifier for clarity
# Requires Python>=3.12